+++ src/main/java/com/apps/pochak/elastic/like/LikeElasticRepository.java
@@ -1,0 +1,6 @@
+package com.apps.pochak.elastic.like;
+
+import com.apps.pochak.elastic.repository.ElasticsearchRepository;
+
+public interface LikeElasticRepository extends ElasticsearchRepository<LikeDocument, String> {
+}
