+++ src/main/java/com/apps/pochak/elastic/member/MemberElasticRepository.java
@@ -1,0 +1,6 @@
+package com.apps.pochak.elastic.member;
+
+import com.apps.pochak.elastic.repository.ElasticsearchRepository;
+
+public interface MemberElasticRepository extends ElasticsearchRepository<MemberDocument, String> {
+}
