--- src/main/java/com/apps/pochak/like/domain/repository/LikeElasticRepository.java
@@ -1,7 +1,0 @@
-package com.apps.pochak.like.domain.repository;
-
-import com.apps.pochak.global.elasticsearch.repository.ElasticsearchRepository;
-import com.apps.pochak.like.domain.LikeDocument;
-
-public interface LikeElasticRepository extends ElasticsearchRepository<LikeDocument, String> {
-}
