--- src/main/java/com/apps/pochak/member/domain/repository/MemberElasticRepository.java
@@ -1,7 +1,0 @@
-package com.apps.pochak.member.domain.repository;
-
-import com.apps.pochak.global.elasticsearch.repository.ElasticsearchRepository;
-import com.apps.pochak.member.domain.MemberDocument;
-
-public interface MemberElasticRepository extends ElasticsearchRepository<MemberDocument, String> {
-}
